ref(all):add permission check point in services

+++ message/src/main/java/com/qzero/bt/message/service/ChatSessionService.java
@@ -1,6 +1,6 @@
 package com.qzero.bt.message.service;
 
-import com.qzero.bt.common.authorize.dao.UserInfoDao;
+import com.qzero.bt.common.authorize.dao.UserInfoRepository;
 import com.qzero.bt.common.exception.ErrorCodeList;
 import com.qzero.bt.common.exception.ResponsiveException;
 import com.qzero.bt.message.data.session.*;
@@ -28,25 +28,38 @@ public class ChatSessionService {
     private ChatMemberDao memberRepository;
 
     @Autowired
-    private UserInfoDao userInfoDao;
+    private UserInfoRepository userInfoRepository;
 
     @Autowired
     private ChatSessionParameterDao parameterDao;
 
-    public void createSession(ChatSession chatSession){
+    @Autowired
+    private SessionParameterCheckManager parameterCheckManager;
+
+    public void createSession(ChatSession chatSession) throws ResponsiveException {
+        if(!parameterCheckManager.checkCompulsoryParameter(chatSession))
+            throw new ResponsiveException(ErrorCodeList.CODE_BAD_REQUEST_PARAMETER,"Session parameter is illegal");
+
         sessionRepository.save(chatSession);
     }
 
-    public void deleteSession(String sessionId){
+    public void deleteSession(String sessionId,String operatorName) throws ResponsiveException {
+        if(!memberRepository.existsBySessionIdAndUserNameAndLevelIsGreaterThanEqual(sessionId,operatorName,ChatMember.LEVEL_OWNER))
+            throw new ResponsiveException(ErrorCodeList.CODE_PERMISSION_DENIED,"You are not the session owner");
         sessionRepository.deleteById(sessionId);
     }
 
-    public ChatSession findSession(String sessionId){
+    public ChatSession findSession(String sessionId,String operatorName) throws ResponsiveException {
+        if(!memberRepository.existsBySessionIdAndUserName(sessionId,operatorName))
+            throw new ResponsiveException(ErrorCodeList.CODE_PERMISSION_DENIED,"You are not one of the members");
         return sessionRepository.getOne(sessionId);
     }
 
-    public void addChatMember(ChatMember chatMember) throws ResponsiveException {
-        if(userInfoDao.getUserInfo(chatMember.getUserName())==null)
+    public void addChatMember(ChatMember chatMember,String operatorName) throws ResponsiveException {
+        if(!isOperator(chatMember.getSessionId(),operatorName))
+            throw new ResponsiveException(ErrorCodeList.CODE_PERMISSION_DENIED,"You are not the session operator");
+
+        if(!userInfoRepository.existsById(chatMember.getUserName()))
             throw new ResponsiveException(ErrorCodeList.CODE_MISSING_RESOURCE,"Target user does not exist");
 
         if(!sessionRepository.existsById(chatMember.getSessionId()))
@@ -58,11 +71,17 @@ public class ChatSessionService {
         memberRepository.save(chatMember);
     }
 
-    public void removeChatMember(ChatMember chatMember){
+    public void removeChatMember(ChatMember chatMember,String operatorName) throws ResponsiveException {
+        //You can quit a session with this method
+        if(!chatMember.equals(operatorName) && !isOperator(chatMember.getSessionId(),operatorName))
+            throw new ResponsiveException(ErrorCodeList.CODE_PERMISSION_DENIED,"You are not the session operator");
         memberRepository.deleteBySessionIdAndUserName(chatMember.getSessionId(),chatMember.getUserName());
     }
 
-    public void updateChatMemberLevel(ChatMember chatMember) throws ResponsiveException {
+    public void updateChatMemberLevel(ChatMember chatMember,String operatorName) throws ResponsiveException {
+        if(!isOperator(chatMember.getSessionId(),operatorName))
+            throw new ResponsiveException(ErrorCodeList.CODE_PERMISSION_DENIED,"You are not the session operator");
+
         ChatMember origin=memberRepository.findBySessionIdAndUserName(chatMember.getSessionId(),chatMember.getUserName());
         if(origin==null)
             throw new ResponsiveException(ErrorCodeList.CODE_MISSING_RESOURCE,"Chat member does not exist");
@@ -76,10 +95,6 @@ public class ChatSessionService {
         origin.setLevel(chatMember.getLevel());
 
         memberRepository.save(origin);
-    }
-
-    public boolean isMemberIn(String sessionId,String userName){
-       return memberRepository.existsBySessionIdAndUserName(sessionId,userName);
     }
 
     public boolean isOperator(String sessionId,String userName){
@@ -103,20 +118,33 @@ public class ChatSessionService {
         return result;
     }
 
-    public void updateSessionParameter(ChatSession session,String parameterName,String parameterValue){
-        ChatSessionParameter parameter=parameterDao.findBySessionIdAndAndParameterName(session.getSessionId(),parameterName);
+    public void updateSessionParameter(String sessionId,String parameterName,String parameterValue,String operatorName) throws IllegalAccessException, ResponsiveException {
+        if(!parameterCheckManager.checkOperationPermission(sessionId,parameterName,
+                operatorName, SessionParameterCheckManager.Operation.OPERATION_UPDATE))
+            throw new ResponsiveException(ErrorCodeList.CODE_PERMISSION_DENIED,"You have no access to the parameter named "+parameterName);
+
+        ChatSessionParameter parameter=parameterDao.findBySessionIdAndAndParameterName(sessionId,parameterName);
         parameter.setParameterValue(parameterValue);
         parameterDao.save(parameter);
     }
 
-    public void deleteSessionParameter(ChatSession session,String parameterName){
-        ChatSessionParameter parameter=parameterDao.findBySessionIdAndAndParameterName(session.getSessionId(),parameterName);
+    public void deleteSessionParameter(String sessionId,String parameterName,String operatorName) throws ResponsiveException, IllegalAccessException {
+        if(!parameterCheckManager.checkOperationPermission(sessionId,parameterName,
+                operatorName, SessionParameterCheckManager.Operation.OPERATION_DELETE))
+            throw new ResponsiveException(ErrorCodeList.CODE_PERMISSION_DENIED,"You have no access to the parameter named "+parameterName);
+
+        ChatSessionParameter parameter=parameterDao.findBySessionIdAndAndParameterName(sessionId,parameterName);
         parameterDao.delete(parameter);
     }
 
-    public void addSessionParameter(ChatSession session,String parameterName,String parameterValue){
-        ChatSessionParameter parameter=new ChatSessionParameter(null,session.getSessionId(),parameterName,parameterValue);
-        parameter.setSessionId(session.getSessionId());
+    public void addSessionParameter(String sessionId,String parameterName,String parameterValue,String operatorName) throws ResponsiveException, IllegalAccessException {
+        if(!parameterCheckManager.checkOperationPermission(sessionId,parameterName,
+                operatorName, SessionParameterCheckManager.Operation.OPERATION_INSERT))
+            throw new ResponsiveException(ErrorCodeList.CODE_PERMISSION_DENIED,"You have no access to the parameter named "+parameterName);
+
+
+        ChatSessionParameter parameter=new ChatSessionParameter(null,sessionId,parameterName,parameterValue);
+        parameter.setSessionId(sessionId);
         parameterDao.save(parameter);
     }
