+++ message/src/main/java/com/qzero/bt/message/service/MessageService.java
@@ -1,5 +1,7 @@
 package com.qzero.bt.message.service;
 
+import com.qzero.bt.common.exception.ErrorCodeList;
+import com.qzero.bt.common.exception.ResponsiveException;
 import com.qzero.bt.message.data.session.ChatMemberDao;
 import com.qzero.bt.message.data.session.ChatSessionDao;
 import com.qzero.bt.message.data.message.MessageDao;
@@ -17,29 +19,55 @@ public class MessageService {
     @Autowired
     private MessageDao messageDao;
 
+    @Autowired
+    private ChatMemberDao memberDao;
+
     public void saveMessage(ChatMessage chatMessage) throws Exception {
+        if(!memberDao.existsBySessionIdAndUserName(chatMessage.getSessionId(), chatMessage.getSenderUserName()))
+            throw new ResponsiveException(ErrorCodeList.CODE_PERMISSION_DENIED,"You are not one of the session members");
+
         messageDao.save(chatMessage);
     }
 
-    public ChatMessage getMessage(String messageId) throws Exception {
-        return messageDao.getOne(messageId);
+    public ChatMessage getMessage(String messageId,String operatorName) throws Exception {
+        ChatMessage message=messageDao.getOne(messageId);
+
+        if(!memberDao.existsBySessionIdAndUserName(message.getSessionId(), operatorName))
+            throw new ResponsiveException(ErrorCodeList.CODE_PERMISSION_DENIED,"You are not one of the session members");
+
+        return message;
     }
 
-    public void deleteMessage(String messageId) throws Exception {
+    public void deleteMessage(String messageId,String operatorName) throws Exception {
+        ChatMessage message=messageDao.getOne(messageId);
+
+        if(!memberDao.existsBySessionIdAndUserName(message.getSessionId(), operatorName))
+            throw new ResponsiveException(ErrorCodeList.CODE_PERMISSION_DENIED,"You are not one of the session members");
+
         messageDao.deleteById(messageId);
     }
 
-    public void updateMessageStatus(String messageId,String newStatus) throws Exception {
+    public void updateMessageStatus(String messageId,String newStatus,String operatorName) throws Exception {
         ChatMessage message=messageDao.getOne(messageId);
+
+        if(!memberDao.existsBySessionIdAndUserName(message.getSessionId(), operatorName))
+            throw new ResponsiveException(ErrorCodeList.CODE_PERMISSION_DENIED,"You are not one of the session members");
+
         message.setMessageStatus(newStatus);
         messageDao.save(message);
     }
 
-    public List<ChatMessage> getAllMessages(String sessionId) throws Exception {
+    public List<ChatMessage> getAllMessages(String sessionId,String operatorName) throws Exception {
+        if(!memberDao.existsBySessionIdAndUserName(sessionId, operatorName))
+            throw new ResponsiveException(ErrorCodeList.CODE_PERMISSION_DENIED,"You are not one of the session members");
+
         return messageDao.getMessagesBySessionId(sessionId);
     }
 
-    public void deleteAllMessageBySessionId(String sessionId) throws Exception{
+    public void deleteAllMessageBySessionId(String sessionId,String operatorName) throws Exception{
+        if(!memberDao.existsBySessionIdAndUserName(sessionId, operatorName))
+            throw new ResponsiveException(ErrorCodeList.CODE_PERMISSION_DENIED,"You are not one of the session members");
+
         messageDao.deleteAllMessagesBySessionId(sessionId);
     }
